+++ se_project_react/README.md
@@ -7,3 +7,49 @@ The idea of the application is pretty simple - we make a call to an API, which t
 ## Links
 
 - [Figma Design](https://www.figma.com/file/DTojSwldenF9UPKQZd6RRb/Sprint-10%3A-WTWR)
+
+### Component Structure
+
+The project includes a components directory with the following components:
+
+- App.js
+- Footer.js
+- Header.js
+- ItemCard.js
+- ItemModal.js
+- Main.js
+- ModalWithForm.js
+- WeatherCard.js
+
+## software to install
+
+- nvm
+- npm
+
+## npm commands
+
+### `npm start`
+
+Runs the app in the development mode.\
+Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
+
+### `npm test`
+
+Launches the test runner in the interactive watch mode.\
+
+### `npm run build`
+
+Builds the app for production to the `build` folder.\
+It correctly bundles React in production mode and optimizes the build for the best performance.
+
+The build is minified and the filenames include the hashes.\
+Your app is ready to be deployed!
+
+### `npm run eject`
+
+## how to run the app
+
+- cd to the project root
+- open a terminal type `nvm use`
+- type `npm install`
+- type `npm run build`
