+++ se_project_react/src/App/App.css
@@ -1,0 +1,24 @@
+.weather_selector:checked {
+  background-color: black;
+}
+.modal__input-label {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+}
+.modal__input {
+  padding-top: 5px;
+  padding-bottom: 5px;
+  border: none;
+  border-bottom: 1px solid black;
+}
+.modal__buttons {
+  justify-content: space-between;
+}
+/* .modal__overlay {
+  opacity: 0.5;
+} */
+.input__button {
+  margin-right: 10px;
+  margin-top: 8px;
+}
