+++ se_project_react/src/Footer/Footer.css
@@ -8,9 +8,6 @@
   /*box-sizing: border-box;*/
   /*outline: red solid 1px;*/
   /*border: orange 1px solid;*/
-  max-width: 1358px;
-  max-height: 20px;
-  left: 40px;
   margin-top: 20px;
   padding: 0 20px;
 }
