+++ se_project_react/src/ItemCard/ItemCard.css
@@ -1,8 +1,4 @@
-.card_section {
-  margin-top: 20px;
-}
-
-.card_name {
+.card__name {
   text-align: center;
   font-weight: 500;
   line-height: 20px;
@@ -12,15 +8,12 @@
   transform: translate(-50%, -50%);
 }
 
-.card_image {
+.card__image {
   width: 328px;
   height: 328px;
   border-radius: 10px;
 }
-.card_items {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-  justify-content: center;
-  margin-top: 20px;
+
+.card {
+  position: relative;
 }
