+++ se_project_react/src/Weather/WeatherCard.css
@@ -1,19 +1,14 @@
 .weather {
   position: relative;
 }
-.weather_info {
+.weather__info {
   position: absolute;
-
   top: 20px;
   left: 16pxs;
   color: white;
   font-size: 36px;
 }
-.weather_image {
+.weather__image {
   width: 100%;
   height: 80px;
-}
-
-.main {
-  padding: 0 20px;
 }
