+++ se_project_react/src/utils/constants.js
@@ -39,60 +39,41 @@ const defaultClothingItems = [
 
 export default defaultClothingItems;
 
-export const weatherDescriptions = {
-  200: "Thunderstorm thunderstorm with light rain",
-  201: "Thunderstorm thunderstorm with rain",
-  202: "Thunderstorm thunderstorm with heavy rain",
-  210: "Thunderstorm light thunderstorm",
-  211: "Thunderstorm thunderstorm",
-  212: "Thunderstorm heavy thunderstorm",
-  221: "Thunderstorm ragged thunderstorm",
-  230: "Thunderstorm thunderstorm with light drizzle",
-  231: "Thunderstorm thunderstorm with drizzle",
-  232: "Thunderstorm thunderstorm with heavy drizzle",
-  300: "Drizzle light intensity drizzle",
-  301: "Drizzle drizzle",
-  302: "Drizzle heavy intensity drizzle",
-  310: "Drizzle light intensity drizzle rain",
-  311: "Drizzle drizzle rain",
-  312: "Drizzle heavy intensity drizzle rain",
-  313: "Drizzle shower rain and drizzle",
-  314: "Drizzle heavy shower rain and drizzle",
-  321: "Drizzle shower drizzle",
-  500: "Rain light rain",
-  501: "Rain moderate rain",
-  502: "Rain heavy intensity rain",
-  503: "Rain very heavy rain",
-  504: "Rain extreme rain",
-  511: "Rain freezing rain",
-  520: "Rain light intensity shower rain",
-  521: "Rain shower rain",
-  522: "Rain heavy intensity shower rain",
-  531: "Rain ragged shower rain",
-  600: "Snow light snow",
-  601: "Snow snow",
-  602: "Snow heavy snow",
-  611: "Snow sleet",
-  612: "Snow light shower sleet",
-  613: "Snow shower sleet",
-  615: "Snow light rain and snow",
-  616: "Snow rain and snow",
-  620: "Snow light shower snow",
-  621: "Snow shower snow",
-  622: "Snow heavy shower snow",
-  701: "Mist mist",
-  711: "Smoke smoke",
-  721: "Haze haze",
-  731: "Dust sand/dust whirls",
-  741: "Fog fog",
-  751: "Sand sand",
-  761: "Dust dust",
-  762: "Ash volcanic ash",
-  771: "Squall squalls",
-  781: "Tornado tornado",
-  800: "Clear clear sky",
-  801: "Clouds few clouds: 11-25%",
-  802: "Clouds scattered clouds: 25-50%",
-  803: "Clouds broken clouds: 51-84%",
-  804: "Clouds overcast clouds: 85-100%",
-};
+export const WeatherOptions = [
+  {
+    url: require("../images/day/cloudy.svg").default,
+    day: true,
+    type: "cloudy",
+  },
+  { url: require("../images/day/fog.svg").default, day: true, type: "fog" },
+  { url: require("../images/day/rain.svg").default, day: true, type: "rain" },
+  { url: require("../images/day/snow.svg").default, day: true, type: "snow" },
+  { url: require("../images/day/storm.svg").default, day: true, type: "storm" },
+  { url: require("../images/day/sunny.svg").default, day: true, type: "sunny" },
+  {
+    url: require("../images/night/clear_sunny.svg").default,
+    day: false,
+    type: "clear_sunny",
+  },
+  {
+    url: require("../images/night/cloudy.svg").default,
+    day: false,
+    type: "cloudy",
+  },
+  { url: require("../images/night/fog.svg").default, day: false, type: "fog" },
+  {
+    url: require("../images/night/rain.svg").default,
+    day: false,
+    type: "rain",
+  },
+  {
+    url: require("../images/night/snow.svg").default,
+    day: false,
+    type: "snow",
+  },
+  {
+    url: require("../images/night/storm.svg").default,
+    day: false,
+    type: "storm",
+  },
+];
